requirements.txt
keyboard==0.13.5
pyperclip==1.8.2
google-generativeai==0.3.2
python-dotenv==1.0.0